adding comments to Game packages classes and moving classes in Game package to separate packages
