+++ src/Game/GameState.java
@@ -1,5 +1,8 @@
 package Game;
 
+/*
+ * This is used by the ScreenCoordinator class to determine which "state" the game is currently in
+ */
 public enum GameState {
     MENU, LEVEL, CREDITS
 }
