--- src/Game/LevelState.java
@@ -1,5 +1,0 @@
-package Game;
-
-public enum LevelState {
-    RUNNING, LEVEL_COMPLETED, PLAYER_DEAD, LEVEL_WIN_MESSAGE, LEVEL_LOSE_MESSAGE, DONE
-}
