+++ src/Game/ScreenCoordinator.java
@@ -4,9 +4,19 @@ import Engine.DefaultScreen;
 import Engine.GraphicsHandler;
 import Engine.Keyboard;
 import Engine.Screen;
+import Screens.CreditsScreen;
+import Screens.MenuScreen;
+import Screens.PlayLevelScreen;
 
+/*
+ * Based on the current game state, this class determines which Screen should be shown
+ * There can only be one "currentScreen", although screens can have "nested" screens
+ */
 public class ScreenCoordinator extends Screen {
+	// currently shown Screen
 	protected Screen currentScreen = new DefaultScreen();
+
+	// keep track of gameState so ScreenCoordinator knows which Screen to show
 	protected GameState gameState;
 	protected GameState previousGameState;
 
@@ -14,18 +24,22 @@ public class ScreenCoordinator extends Screen {
 		return gameState;
 	}
 
+	// Other Screens can set the gameState of this class to force it to change the currentScreen
 	public void setGameState(GameState gameState) {
 		this.gameState = gameState;
 	}
 
 	@Override
 	public void initialize() {
+		// start game off with Menu Screen
 		gameState = GameState.MENU;
 	}
 
 	@Override
 	public void update(Keyboard keyboard) {
 		do {
+			// if previousGameState does not equal gameState, it means there was a change in gameState
+			// this triggers ScreenCoordinator to bring up a new Screen based on what the gameState is
 			if (previousGameState != gameState) {
 				switch(gameState) {
 					case MENU:
@@ -42,12 +56,14 @@ public class ScreenCoordinator extends Screen {
 			}
 			previousGameState = gameState;
 
+			// call the update method for the currentScreen
 			currentScreen.update(keyboard);
 		} while (previousGameState != gameState);
 	}
 
 	@Override
 	public void draw(GraphicsHandler graphicsHandler) {
+		// call the draw method for the currentScreen
 		currentScreen.draw(graphicsHandler);
 	}
 }
