+++ src/Scene/LevelState.java
@@ -1,0 +1,5 @@
+package Scene;
+
+public enum LevelState {
+    RUNNING, LEVEL_COMPLETED, PLAYER_DEAD, LEVEL_WIN_MESSAGE, LEVEL_LOSE_MESSAGE
+}
